Create infrastructure for nested query building
Solarium_Client_Builder is an abstract class that encapsulates helper functions
that both the request builder and nested query builder should be able to use.
The abstract function that both classes define is build(), which generates a request
for the request builder, and a nested query string for the string builder.