Make sure floats use "." decimal separator
The current locale applies when casting floats to strings. For example, with the `de_DE` locale, a boost value of 12.34 will be case to `"12,34"`. This causes a downstream exception in Solr because it does not understand this format.

`number_format` is a bit hacky but is independent of the current locale.