Add support for grouping result using a function
Add functionality that checks whether the result data in
ResponseParser\Component\Grouping contain groups created by
solr parameter `group.func` (an arbitrary Solr-supported
function).

When such field is found, apply the same parsing logic
which is usually used for field-grouping, in order to ensure
the function-grouping is recognized and added to the FieldGroup
result object.